+++ pyengine/Entities/__init__.py
@@ -1,0 +1,2 @@
+from pyengine.Entities.Entity import Entity
+from pyengine.Entities.Tilemap import Tilemap