+++ pyengine/Prefabs/__init__.py
@@ -1,1 +1,0 @@
-from pyengine.Prefabs.Tilemap import Tilemap