+++ pyengine/__init__.py
@@ -1,6 +1,5 @@
 try:
     from pyengine.Window import Window, WindowCallbacks
-    from pyengine.Entity import Entity
     from pyengine.World import World
     from pyengine.Components.PhysicsComponent import CollisionCauses
     from pyengine.Components.ControlComponent import ControlType, Controls, MouseButton
@@ -8,4 +7,4 @@ try:
 except ModuleNotFoundError:
     pass
 
-__version__ = "1.4.2"
+__version__ = "1.5.0"